requirements.txt
requests
prance
openai
pytest
openapi-spec-validator
python-dotenv